+++ backend/.env.development
@@ -2,6 +2,6 @@
 DATABASE_URL="postgres://localhost/books_development"
 SERVE_STATIC_ASSETS="true"
 WEB_SESSIONS_SECRET="b8f07760df659f81e1cb41fd28ede792ddcd843b3ea61c561c2f0a151eb21816"
-ANONYMOUS_SESSIONS_SECRET="c49e1053938f3f9643300d504a07474db456338e9fcae136b276cd02e9e522e8"
+USERS_SESSIONS_SECRET="c49e1053938f3f9643300d504a07474db456338e9fcae136b276cd02e9e522e8"
 AUTH_TOKEN_SECRET="5c1ec52edc7850bafaf26323be8d8a962fa16be35f56a7dffb3513a671a72b09"
 FRONTEND_URL=http://localhost:3000
