+++ backend/.env.test
@@ -2,7 +2,7 @@
 DATABASE_URL="postgres://localhost/books_test"
 SERVE_STATIC_ASSETS="true"
 WEB_SESSIONS_SECRET="83e68857837839326160c52d6f81f85a7f9d9bd5671bfea0bbf48cf0f7b01f87"
-ANONYMOUS_SESSIONS_SECRET="9a477eefd767321ec18990340230966d867d5e394b4d7b059438e1bb03cd1a3d"
+USERS_SESSIONS_SECRET="9a477eefd767321ec18990340230966d867d5e394b4d7b059438e1bb03cd1a3d"
 AUTH_TOKEN_SECRET="5c1ec52edc7850bafaf26323be8d8a962fa16be35f56a7dffb3513a671a72b09"
 
 FRONTEND_URL=http://localhost:3000
