--- backend/apps/anonymous/controllers/graphql/run.rb
@@ -1,12 +1,0 @@
-module Anonymous
-  module Controllers
-    module Graphql
-      class Run
-        include Anonymous::Action
-
-        def call(params)
-        end
-      end
-    end
-  end
-end
