--- backend/apps/anonymous/controllers/oauth/authorize.rb
@@ -1,25 +1,0 @@
-require_relative "client"
-require "controllers/cors"
-
-module Anonymous::Controllers::Oauth
-  class Authorize
-    include Anonymous::Action
-    include Controllers::CORS
-    include Client
-
-    def call(params)
-      session[:state] = SecureRandom.hex(16)
-
-      authorize_url = client.auth_code.authorize_url(
-        redirect_uri: params[:redirect_uri],
-        state: session[:state],
-        scope: "user:email",
-      )
-
-      self.format = :json
-      self.body = {
-        github_authorize_url: authorize_url,
-      }.to_json
-    end
-  end
-end
