--- backend/apps/anonymous/graphql/base_union.rb
@@ -1,6 +1,0 @@
-module Anonymous
-  module GraphQL
-    class BaseUnion < ::GraphQL::Schema::Union
-    end
-  end
-end
