--- backend/apps/anonymous/graphql/mutation_type.rb
@@ -1,11 +1,0 @@
-require_relative 'mutations/users/authenticate'
-
-module Anonymous
-  module GraphQL
-    class MutationType < ::GraphQL::Schema::Object
-      graphql_name "Mutations"
-
-      field :login, mutation: Mutations::Users::Authenticate
-    end
-  end
-end
