--- backend/apps/anonymous/graphql/objects/failed_login_result.rb
@@ -1,11 +1,0 @@
-module Anonymous
-  module GraphQL
-    class FailedLoginResult < ::GraphQL::Schema::Object
-      field :error, String, null: false
-
-      def error
-        object.failure
-      end
-    end
-  end
-end
