--- backend/apps/anonymous/graphql/objects/successful_login_result.rb
@@ -1,16 +1,0 @@
-module Anonymous
-  module GraphQL
-    class SuccessfulLoginResult < ::GraphQL::Schema::Object
-      field :email, String, null: false
-      field :token, String, null: false
-
-      def email
-        object.success[:email]
-      end
-
-      def token
-        object.success[:token]
-      end
-    end
-  end
-end
