--- backend/apps/anonymous/graphql/schema.rb
@@ -1,11 +1,0 @@
-require 'graphql'
-
-require_relative 'mutation_type'
-
-module Anonymous
-  module GraphQL
-    class Schema < ::GraphQL::Schema
-      mutation MutationType
-    end
-  end
-end
