+++ backend/apps/users/container.rb
@@ -1,0 +1,9 @@
+module Users
+  class Container
+    extend Dry::Container::Mixin
+  end
+
+  Import = Dry::AutoInject(Container)
+
+  Container.register(:user_repo, -> { UserRepository.new })
+end
