+++ backend/apps/users/controllers/graphql/run.rb
@@ -1,0 +1,12 @@
+module Users
+  module Controllers
+    module Graphql
+      class Run
+        include Users::Action
+
+        def call(params)
+        end
+      end
+    end
+  end
+end
