+++ backend/apps/users/controllers/oauth/client.rb
@@ -1,0 +1,16 @@
+require 'oauth2'
+
+module Users::Controllers::Oauth
+  module Client
+    def client
+      OAuth2::Client.new(
+        ENV.fetch('OAUTH_CLIENT_ID'),
+        ENV.fetch('OAUTH_CLIENT_SECRET'),
+        authorize_url: '/login/oauth/authorize',
+        token_url: '/login/oauth/access_token',
+        site: 'https://github.com',
+        raise_errors: false,
+      )
+    end
+  end
+end
