+++ backend/apps/users/graphql/base_union.rb
@@ -1,0 +1,6 @@
+module Users
+  module GraphQL
+    class BaseUnion < ::GraphQL::Schema::Union
+    end
+  end
+end
