+++ backend/apps/users/graphql/mutation_type.rb
@@ -1,0 +1,11 @@
+require_relative 'mutations/users/authenticate'
+
+module Users
+  module GraphQL
+    class MutationType < ::GraphQL::Schema::Object
+      graphql_name "Mutations"
+
+      field :login, mutation: Mutations::Users::Authenticate
+    end
+  end
+end
