+++ backend/apps/users/graphql/objects/failed_login_result.rb
@@ -1,0 +1,11 @@
+module Users
+  module GraphQL
+    class FailedLoginResult < ::GraphQL::Schema::Object
+      field :error, String, null: false
+
+      def error
+        object.failure
+      end
+    end
+  end
+end
