+++ backend/apps/users/graphql/schema.rb
@@ -1,0 +1,11 @@
+require 'graphql'
+
+require_relative 'mutation_type'
+
+module Users
+  module GraphQL
+    class Schema < ::GraphQL::Schema
+      mutation MutationType
+    end
+  end
+end
